Add iOS support and fix scale connections
iOS Build Fixes:
- Add Q_OS_IOS guards for QProcess (not available on iOS)
- Compile C++ as Objective-C++ to fix Foundation header conflicts
- Set iOS deployment target to 17.0 (required by Qt 6.10.1)
- Exclude FFmpeg multimedia plugin (use Darwin/AVFoundation)
- Enable iOS file sharing for debug log access
- Add Bluetooth permission request for iOS

Scale Connection Fixes (all platforms):
- Pass full QBluetoothDeviceInfo to transport instead of address string
- iOS doesn't expose MAC addresses, uses UUID instead
- Add getDeviceIdentifier() helper for cross-platform device identification
- Add connectToDevice(QBluetoothDeviceInfo) overload to transport

Other:
- Fix qsizetype to int warnings in screensavervideomanager.h
- Add MissingPermissionsError handling in BLE scan errors
- Enable parallel builds for iOS

+++ src/screensaver/screensavervideomanager.h
@@ -132,7 +132,7 @@ public:
     QString catalogUrl() const { return m_catalogUrl; }
     bool isRefreshing() const { return m_isRefreshing; }
     QDateTime lastUpdatedUtc() const { return m_lastUpdatedUtc; }
-    int itemCount() const { return m_catalog.size(); }
+    int itemCount() const { return static_cast<int>(m_catalog.size()); }
 
     // Category getters
     QVariantList categories() const;
@@ -155,7 +155,7 @@ public:
     QVariantList creditsList() const;
 
     // Personal media getters
-    int personalMediaCount() const { return m_personalCatalog.size(); }
+    int personalMediaCount() const { return static_cast<int>(m_personalCatalog.size()); }
     bool hasPersonalMedia() const { return !m_personalCatalog.isEmpty(); }
     QString personalMediaDirectory() const { return m_cacheDir + "/personal"; }
     bool isPersonalCategory() const { return m_selectedCategoryId == "personal"; }
